# api/docs/swagger.yaml
basePath: /
definitions:
  genproto.CreateMenuRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      restaurant_id:
        type: string
    type: object
  genproto.CreateOrderRequest:
    properties:
      menu_item_id:
        type: string
      quantity:
        type: string
      reservation_id:
        type: string
    type: object
...
  genproto.MenuResponse:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      restaurant_id:
        type: string
    type: object
  genproto.MenusResponse:
    properties:
      menus:
        items:
          $ref: '#/definitions/genproto.MenuResponse'
        type: array
    type: object
  genproto.OrderResponse:
    properties:
      menu_item_id:
        type: string
      quantity:
        type: string
      reservation_id:
        type: string
    type: object
  genproto.OrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/genproto.OrderResponse'
        type: array
    type: object
...
  genproto.UpdateMenuRequest:
    properties:
      Id:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      restaurant_id:
        type: string
    type: object
  genproto.UpdateOrderRequest:
    properties:
      Id:
        type: string
      menu_item_id:
        type: string
      quantity:
        type: string
      reservation_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: APi service
  version: "1.0"
paths:
  /api/menu/create:
    post:
      consumes:
      - application/json
      description: Create a new menu item
      parameters:
      - description: Create Menu
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Menu
      tags:
      - Menu
  /api/menu/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing menu item
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Menu
      tags:
      - Menu
  /api/menu/get_all:
    get:
      consumes:
      - application/json
      description: Get all menu items
      parameters:
      - in: query
        name: description
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: price
        type: number
      - in: query
        name: restaurant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MenusResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get All Menus
      tags:
      - Menu
  /api/menu/get_id/{id}:
    get:
      consumes:
      - application/json
      description: Get a menu item by its ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MenuResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Menu by ID
      tags:
      - Menu
  /api/menu/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing menu item
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Menu
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Menu
      tags:
      - Menu
  /api/order/create:
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Create Order
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Order
      tags:
      - Order
  /api/order/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Order
      tags:
      - Order
  /api/order/get_all:
    get:
      consumes:
      - application/json
      description: Get all orders
      parameters:
      - in: query
        name: menu_item_id
        type: string
      - in: query
        name: quantity
        type: string
      - in: query
        name: reservation_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.OrdersResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get All Orders
      tags:
      - Order
  /api/order/get_id/{id}:
    get:
      consumes:
      - application/json
      description: Get an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Order by ID
      tags:
      - Order
  /api/order/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Order
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Order
      tags:
      - Order
...
swagger: "2.0"